refactor: slim down lambda binaries
