+++ lambda/src/bin/api.rs
@@ -11,7 +11,6 @@
 //! Correlation IDs (UUID v4) are propagated as part of the enqueued task to enable
 //! API→Worker traceability in logs.
 use anyhow::Result;
-use aws_sdk_sqs::Client as SqsClient;
 use hmac::{Hmac, Mac};
 use lambda_runtime::{Error, LambdaEvent, run, service_fn};
 use once_cell::sync::Lazy;
@@ -26,217 +25,16 @@ use uuid::Uuid;
 
 // Import shared modules
 use tldr::SlackBot;
+use tldr::features::api::{
+    build_task_from_view, is_interactive_body, parse_interactive_payload, send_to_sqs, v_str,
+};
 use tldr::{
     Prefill, SlackError, build_tldr_modal, sanitize_custom_prompt,
-    slack_parser::{SlackCommandEvent, decode_url_component, parse_form_data},
+    slack_parser::{SlackCommandEvent, parse_form_data},
     validate_view_submission,
 };
 
 use tldr::core::models::ProcessingTask;
-
-async fn send_to_sqs(task: &ProcessingTask, config: &AppConfig) -> Result<(), SlackError> {
-    // Get queue URL from environment
-    let queue_url = &config.processing_queue_url;
-
-    // Set up AWS SDK with new API
-    let shared_config = aws_config::from_env().load().await;
-    let client = SqsClient::new(&shared_config);
-
-    // Serialize task to JSON string
-    let message_body = serde_json::to_string(task)
-        .map_err(|e| SlackError::ApiError(format!("Failed to serialize task: {}", e)))?;
-
-    // Use the builder pattern correctly for SQS client
-    client
-        .send_message()
-        .queue_url(queue_url)
-        .message_body(message_body)
-        .send()
-        .await
-        .map_err(|e| SlackError::AwsError(format!("Failed to send message to SQS: {}", e)))?;
-
-    Ok(())
-}
-
-/// Detects whether the incoming body is a Slack interactive payload (form-encoded with a `payload=` JSON string)
-fn is_interactive_body(body: &str) -> bool {
-    // Check if body starts with "payload=" or contains "&payload=" to avoid false positives
-    body.starts_with("payload=") || body.contains("&payload=")
-}
-
-/// Parses the interactive `payload` JSON from a form-encoded body
-fn parse_interactive_payload(form_body: &str) -> Result<Value, SlackError> {
-    // Very small parser for key=value&key=value to extract `payload`
-    for pair in form_body.split('&') {
-        if let Some(eq_idx) = pair.find('=') {
-            let key = &pair[..eq_idx];
-            if key == "payload" {
-                let raw_val = &pair[eq_idx + 1..];
-                let decoded = decode_url_component(raw_val).map_err(|e| {
-                    SlackError::ParseError(format!("Failed to decode payload: {}", e))
-                })?;
-                let v: Value = serde_json::from_str(&decoded)
-                    .map_err(|e| SlackError::ParseError(format!("Invalid JSON payload: {}", e)))?;
-                return Ok(v);
-            }
-        }
-    }
-    Err(SlackError::ParseError("Missing payload field".to_string()))
-}
-
-/// Helper: traverse a nested JSON object by path of keys.
-fn v_path<'a>(root: &'a Value, path: &[&str]) -> Option<&'a Value> {
-    let mut cur = root;
-    for key in path {
-        cur = cur.get(*key)?;
-    }
-    Some(cur)
-}
-
-/// Helper: get a nested string value by path.
-fn v_str<'a>(root: &'a Value, path: &[&str]) -> Option<&'a str> {
-    v_path(root, path).and_then(|v| v.as_str())
-}
-
-/// Helper: get a nested array value by path.
-fn v_array<'a>(root: &'a Value, path: &[&str]) -> Option<&'a Vec<Value>> {
-    v_path(root, path).and_then(|v| v.as_array())
-}
-
-/// Build a ProcessingTask from a `view_submission` payload's view.state.values
-///
-/// Parses the Block Kit modal submission data structure:
-/// - view.state.values contains blocks indexed by block_id
-/// - Each block contains action_ids with the user's input
-/// - Extracts conversation, range mode, message count, destination, and custom prompt
-fn build_task_from_view(
-    user_id: &str,
-    view: &Value,
-    correlation_id: String,
-) -> Result<ProcessingTask, SlackError> {
-    // Ensure view.state.values exists (Block Kit requirement)
-    let _ = v_path(view, &["state", "values"]) // only for presence check
-        .and_then(|v| v.as_object())
-        .ok_or_else(|| SlackError::ParseError("view.state.values missing".to_string()))?;
-
-    // Extract selected conversation from the conversations_select block
-    // Path: state.values.conv.conv_id.selected_conversation
-    let channel_id = v_str(
-        view,
-        &[
-            "state",
-            "values",
-            "conv",    // block_id
-            "conv_id", // action_id
-            "selected_conversation",
-        ],
-    )
-    .unwrap_or("")
-    .to_string();
-
-    // Extract range mode from radio button selection
-    // Path: state.values.range.mode.selected_option.value
-    let mode = v_str(
-        view,
-        &[
-            "state",
-            "values",
-            "range", // block_id
-            "mode",  // action_id
-            "selected_option",
-            "value",
-        ],
-    )
-    .unwrap_or("unread_since_last_run");
-
-    // Extract optional message count for "last N messages" mode
-    // Path: state.values.lastn.n.value
-    let message_count = v_str(view, &["state", "values", "lastn", "n", "value"])
-        .and_then(|s| s.parse::<u32>().ok());
-
-    // Parse destination checkboxes to determine output routing
-    // Path: state.values.dest.dest_flags.selected_options[]
-    let mut dest_canvas = false;
-    let mut dest_dm = false;
-    let mut dest_public_post = false;
-
-    if let Some(selected) = v_array(
-        view,
-        &["state", "values", "dest", "dest_flags", "selected_options"],
-    ) {
-        for opt in selected {
-            if let Some(val) = opt.get("value").and_then(|s| s.as_str()) {
-                match val {
-                    "canvas" => dest_canvas = true,
-                    "dm" => dest_dm = true,
-                    "public_post" => dest_public_post = true,
-                    _ => {}
-                }
-            }
-        }
-    }
-
-    // visible flag is true if posting publicly
-    let visible = dest_public_post;
-
-    // Extract and sanitize custom prompt if provided
-    // Path: state.values.style.custom.value
-    let custom_prompt = v_str(view, &["state", "values", "style", "custom", "value"])
-        .map(|s| s.to_string())
-        .and_then(|raw| sanitize_custom_prompt(&raw).ok());
-
-    // Determine effective message count based on mode
-    // Only use explicit count for "last_n" mode; otherwise None = unread messages
-    let effective_count = if mode == "last_n" {
-        message_count
-    } else {
-        None
-    };
-
-    // Build text representation of UI selections for attribution when posting publicly
-    let mut text_parts = Vec::new();
-
-    // Add count if specified
-    if let Some(count) = effective_count {
-        text_parts.push(format!("count={}", count));
-    }
-
-    // Add custom prompt if present
-    if let Some(ref prompt) = custom_prompt {
-        // Truncate long prompts for display (max 100 chars)
-        // Use char-boundary-safe truncation to avoid panic
-        let display_prompt = if prompt.chars().count() > 100 {
-            let truncated: String = prompt.chars().take(97).collect();
-            format!("custom=\"{}...\"", truncated)
-        } else {
-            format!("custom=\"{}\"", prompt)
-        };
-        text_parts.push(display_prompt);
-    }
-
-    // Add visibility flags
-    if dest_public_post {
-        text_parts.push("--visible".to_string());
-    }
-
-    // Combine into text field for attribution
-    let text = text_parts.join(" ");
-
-    Ok(ProcessingTask {
-        correlation_id,
-        user_id: user_id.to_string(),
-        channel_id,
-        response_url: None,
-        text,
-        message_count: effective_count,
-        target_channel_id: None,
-        custom_prompt,
-        visible,
-        dest_canvas,
-        dest_dm,
-        dest_public_post,
-    })
-}
 
 #[allow(dead_code)]
 async fn get_latest_message_ts(
