+++ lambda/src/lib.rs
@@ -68,6 +68,7 @@ pub mod clients;
 pub mod core;
 pub mod domains;
 pub mod errors;
+pub mod features;
 pub mod formatting;
 pub mod prompt;
 pub mod response;
